+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/old/BlueLeftAuto.java
@@ -1,6 +1,7 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.auto.old;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
+import com.acmerobotics.roadrunner.geometry.Vector2d;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
@@ -9,12 +10,14 @@ import com.qualcomm.robotcore.hardware.DistanceSensor;
 import com.qualcomm.robotcore.hardware.Servo;
 
 import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
+import org.firstinspires.ftc.teamcode.functions.ArmRamp;
+import org.firstinspires.ftc.teamcode.functions.Constants;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 
-@Autonomous(name="Blue Right Auto", group = "Centerstage Autonomous", preselectTeleOp = "RobotController")
+@Autonomous(name="Blue Left Auto", group = "Centerstage Autonomous", preselectTeleOp = "RobotController")
 @Disabled
-public class BlueRightAuto extends LinearOpMode {
+public class BlueLeftAuto extends LinearOpMode {
     //Sets up motors and variables
     DistanceSensor leftSensor;
     DistanceSensor rightSensor;
@@ -31,23 +34,23 @@ public class BlueRightAuto extends LinearOpMode {
     Servo arm;
     double slidePosError;
     double slidePower;
-    double waitTime;
+
     Servo droneLauncher;
 
     @Override
     public void runOpMode() throws InterruptedException
     {
-        //Sets sensors
+        //Sets up sensors
         leftSensor = hardwareMap.get(DistanceSensor.class, "checkLeft");
         rightSensor = hardwareMap.get(DistanceSensor.class, "checkRight");
-        //Sets servos
+        //Sets up servos
         leftGripper = hardwareMap.servo.get("leftGripperServo");
         rightGripper = hardwareMap.servo.get("rightGripperServo");
-        //Sets slide motors
+
         leftGripper.setPosition(Constants.GRIPPER_LEFT_CLOSE_POSITION);
         rightGripper.setPosition(Constants.GRIPPER_RIGHT_CLOSE_POSITION);
         arm = hardwareMap.servo.get("armServo");
-
+        //Sets up slides
         slideMotor1 = hardwareMap.dcMotor.get("leftSlideMotor");
         slideMotor2 = hardwareMap.dcMotor.get("rightSlideMotor");
 
@@ -64,33 +67,29 @@ public class BlueRightAuto extends LinearOpMode {
         droneLauncher = hardwareMap.servo.get("droneLaunchServo");
         droneLauncher.setPosition(Constants.DRONE_START_POSITION);
 
-        //Sets up dead wheels
         SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
-        Pose2d startPose = new Pose2d(-38.25, 63.75, Math.toRadians(270));
+        Pose2d startPose = new Pose2d(14.5, 63.75, Math.toRadians(-90));
         drive.setPoseEstimate(startPose);
 
         waitForStart();
 
         TrajectorySequence approachSpikeMarks = drive.trajectorySequenceBuilder(startPose)
                 //Move into spike mark area
-                .splineToLinearHeading(new Pose2d(-35, 30.25, startPose.getHeading()), startPose.getHeading())
+                .splineToLinearHeading(new Pose2d(11.75, 30.25, startPose.getHeading()), startPose.getHeading())
                 //Determine where prop is located
                 .addDisplacementMarker(() -> {
                     if (leftSensor.getDistance(DistanceUnit.INCH) < 5) {
-                        selectPose = new Pose2d(-32, 31.25, Math.toRadians(0));
+                        selectPose = new Pose2d(15.75, 31.25, Math.toRadians(0));
                         selectTurn = 90;
-                        boardOffset = 5.75;
-                        waitTime = 0;
+                        boardOffset = 4.25;
                     } else if (rightSensor.getDistance(DistanceUnit.INCH) < 5) {
-                        selectPose = new Pose2d(-39, 31.25,Math.toRadians(180));
+                        selectPose = new Pose2d(7.75, 31.25, Math.toRadians(180));
                         selectTurn = -90;
-                        boardOffset = -5.75;
-                        waitTime = 3.25;
+                        boardOffset = -4.25;
                     } else {
-                        selectPose = new Pose2d(-35.5, 31.5, Math.toRadians(270));
+                        selectPose = new Pose2d(11.75, 31.5, Math.toRadians(-90));
                         selectTurn = 0;
-                        boardOffset = 0;
-                        waitTime = 2;
+                        boardOffset = 1.5;
                     }
                 })
                 .build();
@@ -108,29 +107,17 @@ public class BlueRightAuto extends LinearOpMode {
                 //Back away from purple pixel
                 .back(4.5)
                 //Move back to safe turning position
-                .lineToLinearHeading(new Pose2d(-37, 48, selectPose.getHeading()))
+                .lineToLinearHeading(new Pose2d(13.75, 50, selectPose.getHeading()))
                 //Turn towards back wall
-                .turn(Math.toRadians(180.001)-selectPose.getHeading())
-                //Drive towards back wall
-                .lineToLinearHeading(new Pose2d(-58,48, Math.toRadians(180)))
-                //Turn towards opposing side
-                .turn(Math.toRadians(90))
-                //Drive towards opposing side
-                .lineToLinearHeading(new Pose2d(-58,9, Math.toRadians(270)))
-                //Turn towards center of field
-                .turn(Math.toRadians(90))
-                //Drive towards center of field
-                .lineToLinearHeading(new Pose2d(15,9,0))
-                //Wait for teammate to finish
-                .waitSeconds(waitTime)
-                //Drive to board
-                .splineToLinearHeading(new Pose2d(39,38.5 + boardOffset,0),Math.toRadians(90))
+                .turn(Math.toRadians(0.001)-selectPose.getHeading())
+                //Drives to board
+                .splineToLinearHeading(new Pose2d(39,37 + boardOffset,0),Math.toRadians(-90))
                 .build();
 
         drive.followTrajectorySequence(placePurplePixel);
 
         //Raise slide and arm
-        slidePos = 600;
+        slidePos = 320;
         armUp = true;
         while ((slideMotor1.getCurrentPosition() < slidePos - 30 || rampPos < Constants.ARM_UP_POS - 0.001) && opModeIsActive()) {
             arm.setPosition(ArmRamp.Ramp(Constants.ARM_DOWN_POS, Constants.ARM_UP_POS, armUp));
@@ -140,6 +127,7 @@ public class BlueRightAuto extends LinearOpMode {
 
             slideMotor1.setPower(slidePower);
             slideMotor2.setPower(slidePower);
+
             telemetry.addData("Slide Pos", slideMotor1.getCurrentPosition());
             telemetry.addData("Arm Pos", arm.getPosition());
             telemetry.update();
@@ -182,7 +170,15 @@ public class BlueRightAuto extends LinearOpMode {
         slideMotor1.setPower(0);
         slideMotor2.setPower(0);
 
-        Constants.autoEndPose = dropOnBoard.end();
+        TrajectorySequence moveToCorner = drive.trajectorySequenceBuilder(dropOnBoard.end())
+                //Moves to back right corner
+                .lineToConstantHeading(new Vector2d(44.5, 60))
+                //Scoots forward to get out of the way
+                .forward(10)
+                .build();
+        drive.followTrajectorySequence(moveToCorner);
+
+        Constants.autoEndPose = moveToCorner.end();
         Constants.blueAuto = true;
     }
 }
