+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/old/BlueRightAutoVision.java
@@ -1,18 +1,18 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.auto.old;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
-import com.acmerobotics.roadrunner.geometry.Vector2d;
 import com.acmerobotics.roadrunner.trajectory.Trajectory;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DistanceSensor;
 import com.qualcomm.robotcore.hardware.Servo;
 
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
-import org.firstinspires.ftc.robotcore.external.hardware.camera.controls.ExposureControl;
-import org.firstinspires.ftc.robotcore.external.hardware.camera.controls.GainControl;
 import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
+import org.firstinspires.ftc.teamcode.functions.ArmRamp;
+import org.firstinspires.ftc.teamcode.functions.Constants;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 import org.firstinspires.ftc.vision.VisionPortal;
@@ -20,10 +20,10 @@ import org.firstinspires.ftc.vision.apriltag.AprilTagDetection;
 import org.firstinspires.ftc.vision.apriltag.AprilTagProcessor;
 
 import java.util.List;
-import java.util.concurrent.TimeUnit;
 
-@Autonomous(name="Red Left Auto Vision With White", group = "Centerstage Autonomous Red", preselectTeleOp = "RobotController")
-public class RedLeftAutoVisionPickUpWhite extends LinearOpMode {
+@Autonomous(name="Blue Right Auto Vision", group = "Centerstage Autonomous", preselectTeleOp = "RobotController")
+@Disabled
+public class BlueRightAutoVision extends LinearOpMode {
     //Sets up motors and variables
     DistanceSensor leftSensor;
     DistanceSensor rightSensor;
@@ -43,16 +43,13 @@ public class RedLeftAutoVisionPickUpWhite extends LinearOpMode {
     long waitTime;
     int desiredTagId;
     Servo droneLauncher;
-    VisionPortal visionPortal;
 
     @Override
     public void runOpMode() throws InterruptedException
     {
         //Sets up camera
         AprilTagProcessor tagProcessor = AprilTagProcessor.easyCreateWithDefaults();
-        visionPortal = VisionPortal.easyCreateWithDefaults(hardwareMap.get(WebcamName.class, "Webcam 1"), tagProcessor);
-
-        setManualExposure(1, 255);
+        VisionPortal visionPortal = VisionPortal.easyCreateWithDefaults(hardwareMap.get(WebcamName.class, "Webcam 1"), tagProcessor);
 
         //Sets sensors
         leftSensor = hardwareMap.get(DistanceSensor.class, "checkLeft");
@@ -86,30 +83,30 @@ public class RedLeftAutoVisionPickUpWhite extends LinearOpMode {
 
         //Sets up dead wheels
         SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
-        Pose2d startPose = new Pose2d(-38.25, -63.75, Math.toRadians(90));
+        Pose2d startPose = new Pose2d(-38.25, 63.75, Math.toRadians(270));
         drive.setPoseEstimate(startPose);
 
-        Constants.blueAuto = false;
+        Constants.blueAuto = true;
 
         waitForStart();
 
         //Move into spike mark area
         Trajectory approachSpikeMarks = drive.trajectoryBuilder(startPose)
-                .splineToLinearHeading(new Pose2d(-35, -40.25, startPose.getHeading()), startPose.getHeading())
-                .lineToLinearHeading(new Pose2d(-35, -30.25, startPose.getHeading()))
+                .splineToLinearHeading(new Pose2d(-35, 40.25, startPose.getHeading()), startPose.getHeading())
+                .lineToLinearHeading(new Pose2d(-35, 30.25, startPose.getHeading()))
                 .build();
         drive.followTrajectory(approachSpikeMarks);
 
         TrajectorySequence placePurplePixel;
 
         //Detect prop
-        if (leftSensor.getDistance(DistanceUnit.INCH) < 5) {
-            selectPose = new Pose2d(-38.5, -31.25, Math.toRadians(180));
-            selectTurn = 90;
-            boardOffset = 6;
+        if (rightSensor.getDistance(DistanceUnit.INCH) < 5) {
+            selectPose = new Pose2d(-38.5, 31.25, Math.toRadians(180));
+            selectTurn = -90;
+            boardOffset = -6;
             zeOffset = 0;
-            waitTime = 2;
-            desiredTagId = 4;
+            waitTime = 8;
+            desiredTagId = 3;
             placePurplePixel = drive.trajectorySequenceBuilder(approachSpikeMarks.end())
                     //Turn to proper spike mark
                     .turn(Math.toRadians(selectTurn))
@@ -120,15 +117,17 @@ public class RedLeftAutoVisionPickUpWhite extends LinearOpMode {
                         leftGripper.setPosition(Constants.GRIPPER_LEFT_OPEN_POSITION);
                     })
                     .back(5)
-                    .lineToLinearHeading(new Pose2d(-37, -15, Math.toRadians(180)))
+                    .turn(Math.toRadians(90))
+                    .lineToLinearHeading(new Pose2d(-37, 11.5, Math.toRadians(270)))
+                    .turn(Math.toRadians(90))
                     .build();
-        } else if (rightSensor.getDistance(DistanceUnit.INCH) < 5) {
-            selectPose = new Pose2d(-32.5, -31.25,0);
-            selectTurn = -90;
-            boardOffset = -6;
+        } else if (leftSensor.getDistance(DistanceUnit.INCH) < 5) {
+            selectPose = new Pose2d(-32.5, 31.25,0);
+            selectTurn = 90;
+            boardOffset = 6;
             zeOffset = 0;
-            waitTime = 0;
-            desiredTagId = 6;
+            waitTime = 8;
+            desiredTagId = 1;
             placePurplePixel = drive.trajectorySequenceBuilder(approachSpikeMarks.end())
                     //Turn to proper spike mark
                     .turn(Math.toRadians(selectTurn))
@@ -139,22 +138,22 @@ public class RedLeftAutoVisionPickUpWhite extends LinearOpMode {
                         leftGripper.setPosition(Constants.GRIPPER_LEFT_OPEN_POSITION);
                     })
                     .back(7)
-                    .turn(Math.toRadians(90))
-                    .lineToLinearHeading(new Pose2d(-38, -11.5, Math.toRadians(90)))
+                    .turn(Math.toRadians(-90))
+                    .lineToLinearHeading(new Pose2d(-38, 11.5, Math.toRadians(270)))
                     .turn(Math.toRadians(90))
                     .build();
         } else {
-            selectPose = new Pose2d(-35.5, -31.5, Math.toRadians(90));
+            selectPose = new Pose2d(-35.5, 31.5, Math.toRadians(90));
             selectTurn = 180;
             boardOffset = 0;
             zeOffset = 0;
-            waitTime = 1;
-            desiredTagId = 5;
+            waitTime = 10;
+            desiredTagId = 2;
             placePurplePixel = drive.trajectorySequenceBuilder(approachSpikeMarks.end())
                     //Turn to proper spike mark
                     .turn(Math.toRadians(selectTurn))
                     //Drive up to proper spike mark
-                    .lineToLinearHeading(new Pose2d(-40, -16, Math.toRadians(270.001)))
+                    .lineToLinearHeading(new Pose2d(-40, 16, Math.toRadians(90)))
                     .addDisplacementMarker(() -> {
                         leftGripper.setPosition(Constants.GRIPPER_LEFT_OPEN_POSITION);
                     })
@@ -164,97 +163,15 @@ public class RedLeftAutoVisionPickUpWhite extends LinearOpMode {
         }
 
         drive.followTrajectorySequence(placePurplePixel);
-        //TODO: Teleport to pick up white pixel
-        //Raises arm
-        slidePos = Constants.SLIDE_STACK_POS;
-        armUp = false;
-        double lastSlidePower = 0, slidePowerIncr = 0.1;
-        while ((slideMotor1.getCurrentPosition() < slidePos - 30) && opModeIsActive()) {
-            slidePosError = slidePos - slideMotor1.getCurrentPosition();
-            slidePower = slidePosError * 3 / 500;
-            slidePower = Math.min(Math.max(slidePower, -0.6), 0.75);
-
-            if(slidePower - lastSlidePower > slidePowerIncr)
-            {
-                slidePower = lastSlidePower + slidePowerIncr;
-            }
-
-            slideMotor1.setPower(slidePower);
-            slideMotor2.setPower(slidePower);
-            telemetry.addData("Slide Pos", slideMotor1.getCurrentPosition());
-            telemetry.update();
-            lastSlidePower = slideMotor1.getPower();
-        }
-        slideMotor1.setPower(0);
-        slideMotor2.setPower(0);
-      //  sleep(2147483647);
-        TrajectorySequence actuallyPickUpWhitePixel = drive.trajectorySequenceBuilder(placePurplePixel.end())
-                //Drives into pixel stack
-                .splineToLinearHeading(new Pose2d(-60.5, -9.5, Math.toRadians(180)), Math.toRadians(180))
-                .build();
-        drive.followTrajectorySequence(actuallyPickUpWhitePixel);
-
-        //Grabs white pixel
-        leftGripper.setPosition(Constants.GRIPPER_LEFT_CLOSE_POSITION);
-
-        sleep(0);
-
-        TrajectorySequence backUp = drive.trajectorySequenceBuilder(actuallyPickUpWhitePixel.end())
-                //Backs up a little bit to get room
-                .lineToLinearHeading(new Pose2d(-38, -11.5, Math.toRadians(180)))
-                .build();
-        drive.followTrajectorySequence(backUp);
-        //Lowers arm
-        slidePos = 0;
-        armUp = false;
-        lastSlidePower = 0;
-        slidePowerIncr = 0.1;
-        while ((slideMotor1.getCurrentPosition() > slidePos + 30) && opModeIsActive()) {
-            slidePosError = slidePos - slideMotor1.getCurrentPosition();
-            slidePower = slidePosError * 3 / 500;
-            slidePower = Math.min(Math.max(slidePower, -0.6), 0.75);
-
-            if(slidePower - lastSlidePower > slidePowerIncr)
-            {
-                slidePower = lastSlidePower + slidePowerIncr;
-            }
-
-            slideMotor1.setPower(slidePower);
-            slideMotor2.setPower(slidePower);
-            telemetry.addData("Slide Pos", slideMotor1.getCurrentPosition());
-            telemetry.update();
-            lastSlidePower = slideMotor1.getPower();
-        }
-        slideMotor1.setPower(0);
-        slideMotor2.setPower(0);
-
-        //Opens grippers
-        leftGripper.setPosition(Constants.GRIPPER_LEFT_OPEN_POSITION);
-        rightGripper.setPosition(Constants.GRIPPER_RIGHT_OPEN_POSITION);
-
-        TrajectorySequence goForward = drive.trajectorySequenceBuilder(backUp.end())
-                .forward(2)
-                .build();
-        drive.followTrajectorySequence(goForward);
-
-        leftGripper.setPosition(Constants.GRIPPER_LEFT_CLOSE_POSITION);
-        rightGripper.setPosition(Constants.GRIPPER_RIGHT_CLOSE_POSITION);
-        sleep(0);
-
-        TrajectorySequence holdWhitePixel = drive.trajectorySequenceBuilder(goForward.end())
-                .turn(Math.toRadians(-180))
-                .build();
-
-        drive.followTrajectorySequence(holdWhitePixel);
 
         //Waits for teammate to finish
         sleep(waitTime * 1000);
 
-        TrajectorySequence driveToCenterOfField = drive.trajectorySequenceBuilder(holdWhitePixel.end())
+        TrajectorySequence driveToCenterOfField = drive.trajectorySequenceBuilder(placePurplePixel.end())
                 //Drive towards center of field
-                .lineToLinearHeading(new Pose2d(15,-9,0))
+                .lineToLinearHeading(new Pose2d(15,9,0))
                 //Drive to board
-                .splineToLinearHeading(new Pose2d(36,-34 + boardOffset,0), Math.toRadians(-90))
+                .splineToLinearHeading(new Pose2d(36,35.5 + boardOffset,0), Math.toRadians(90))
                 .build();
 
         drive.followTrajectorySequence(driveToCenterOfField);
@@ -262,8 +179,7 @@ public class RedLeftAutoVisionPickUpWhite extends LinearOpMode {
         //Raise slide and arm
         slidePos = 550;
         armUp = true;
-         lastSlidePower = 0;
-         slidePowerIncr = 0.1;
+        double lastSlidePower = 0, slidePowerIncr = 0.1;
         while ((slideMotor1.getCurrentPosition() < slidePos - 30 || ArmRamp.rampPos > Constants.ARM_UP_POS + 0.001) && opModeIsActive()) {
             arm.setPosition(ArmRamp.Ramp(Constants.ARM_DOWN_POS, Constants.ARM_UP_POS, armUp));
             slidePosError = slidePos - slideMotor1.getCurrentPosition();
@@ -280,6 +196,7 @@ public class RedLeftAutoVisionPickUpWhite extends LinearOpMode {
             telemetry.addData("Slide Pos", slideMotor1.getCurrentPosition());
             telemetry.addData("Arm Pos", arm.getPosition());
             telemetry.update();
+
             lastSlidePower = slideMotor1.getPower();
         }
         slideMotor1.setPower(0);
@@ -336,44 +253,17 @@ public class RedLeftAutoVisionPickUpWhite extends LinearOpMode {
             }
         }
 
-        TrajectorySequence moveWhitePixel;
-        if(desiredTagId == 4){
-            //Drives forward to board
-            Trajectory driveToBoard = drive.trajectoryBuilder(driveToCenterOfField.end())
-                    .lineToLinearHeading(new Pose2d(48 - ye, -32.5 - 15 /*# is white pixel offset*/ + boardOffset - ze, Math.toRadians(0 + pe)))
-                    .build();
-            drive.followTrajectory(driveToBoard);
-            leftGripper.setPosition(Constants.GRIPPER_LEFT_OPEN_POSITION);
-
-            moveWhitePixel = drive.trajectorySequenceBuilder(driveToBoard.end())
-                    .back(2)
-                    .lineToConstantHeading(new Vector2d(driveToBoard.end().getX(), driveToBoard.end().getY() + 15))
-                    .build();
-            drive.followTrajectorySequence(moveWhitePixel);
-
-            rightGripper.setPosition(Constants.GRIPPER_RIGHT_OPEN_POSITION);
-        }
-        else {
-            //Drives forward to board
-            TrajectorySequence driveToBoard = drive.trajectorySequenceBuilder(driveToCenterOfField.end())
-                    .lineToLinearHeading(new Pose2d(48 - ye, -32.5 + boardOffset - ze, Math.toRadians(0 + pe)))
-                    .build();
-            drive.followTrajectorySequence(driveToBoard);
-
-            //Opens grippers
-            rightGripper.setPosition(Constants.GRIPPER_RIGHT_OPEN_POSITION);
-            moveWhitePixel = drive.trajectorySequenceBuilder(driveToBoard.end())
-                    .back(2)
-                    .forward(2)
-                    .build();
-            drive.followTrajectorySequence(moveWhitePixel);
-            leftGripper.setPosition(Constants.GRIPPER_LEFT_OPEN_POSITION);
-        }
-
+        //Drives forward to board
+        Trajectory driveToBoard = drive.trajectoryBuilder(driveToCenterOfField.end())
+                .lineToLinearHeading(new Pose2d(49 - ye, 35.5 + boardOffset - ze, Math.toRadians(0 + pe)))
+                .build();
+        drive.followTrajectory(driveToBoard);
+        //Opens gripper
+        rightGripper.setPosition(Constants.GRIPPER_RIGHT_OPEN_POSITION);
         //Waits for pixel to drop
-        sleep(0);
+        sleep(600);
         //Drives back from board
-        Trajectory driveBackFromBoard = drive.trajectoryBuilder(moveWhitePixel.end())
+        Trajectory driveBackFromBoard = drive.trajectoryBuilder(driveToBoard.end())
                 .back(5)
                 .build();
         drive.followTrajectory(driveBackFromBoard);
@@ -402,44 +292,5 @@ public class RedLeftAutoVisionPickUpWhite extends LinearOpMode {
         slideMotor2.setPower(0);
 
         sleep(30000);
-    }
-
-    private boolean    setManualExposure(int exposureMS, int gain) {
-        // Ensure Vision Portal has been setup.
-        if (visionPortal == null) {
-            return false;
-        }
-
-        // Wait for the camera to be open
-        if (visionPortal.getCameraState() != VisionPortal.CameraState.STREAMING) {
-            telemetry.addData("Camera", "Waiting");
-            telemetry.update();
-            while (!isStopRequested() && (visionPortal.getCameraState() != VisionPortal.CameraState.STREAMING)) {
-                sleep(20);
-            }
-            telemetry.addData("Camera", "Ready");
-            telemetry.update();
-        }
-
-        // Set camera controls unless we are stopping.
-        if (!isStopRequested())
-        {
-            // Set exposure.  Make sure we are in Manual Mode for these values to take effect.
-            ExposureControl exposureControl = visionPortal.getCameraControl(ExposureControl.class);
-            if (exposureControl.getMode() != ExposureControl.Mode.Manual) {
-                exposureControl.setMode(ExposureControl.Mode.Manual);
-                sleep(50);
-            }
-            exposureControl.setExposure((long)exposureMS, TimeUnit.MILLISECONDS);
-            sleep(20);
-
-            // Set Gain.
-            GainControl gainControl = visionPortal.getCameraControl(GainControl.class);
-            gainControl.setGain(gain);
-            sleep(20);
-            return (true);
-        } else {
-            return (false);
-        }
     }
 }
