+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/old/BlueRightAuto_Lower.java
@@ -1,20 +1,22 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.auto.old;
 
+import com.acmerobotics.roadrunner.geometry.Pose2d;
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DistanceSensor;
 import com.qualcomm.robotcore.hardware.Servo;
-import com.acmerobotics.roadrunner.geometry.Pose2d;
 
 import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
+import org.firstinspires.ftc.teamcode.functions.ArmRamp;
+import org.firstinspires.ftc.teamcode.functions.Constants;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 
-@Autonomous(name="Old Red Left Auto", group = "Centerstage Autonomous", preselectTeleOp = "RobotController")
+@Autonomous(name="Blue Right Auto Lower", group = "Centerstage Autonomous", preselectTeleOp = "RobotController")
 @Disabled
-public class OldRedLeftAuto extends LinearOpMode {
+public class BlueRightAuto_Lower extends LinearOpMode {
     //Sets up motors and variables
     DistanceSensor leftSensor;
     DistanceSensor rightSensor;
@@ -40,14 +42,14 @@ public class OldRedLeftAuto extends LinearOpMode {
         //Sets sensors
         leftSensor = hardwareMap.get(DistanceSensor.class, "checkLeft");
         rightSensor = hardwareMap.get(DistanceSensor.class, "checkRight");
-        //Sets grippers and other servos
+        //Sets servos
         leftGripper = hardwareMap.servo.get("leftGripperServo");
         rightGripper = hardwareMap.servo.get("rightGripperServo");
-
+        //Sets slide motors
         leftGripper.setPosition(Constants.GRIPPER_LEFT_CLOSE_POSITION);
         rightGripper.setPosition(Constants.GRIPPER_RIGHT_CLOSE_POSITION);
         arm = hardwareMap.servo.get("armServo");
-        //Sets slide motors
+
         slideMotor1 = hardwareMap.dcMotor.get("leftSlideMotor");
         slideMotor2 = hardwareMap.dcMotor.get("rightSlideMotor");
 
@@ -63,32 +65,33 @@ public class OldRedLeftAuto extends LinearOpMode {
 
         droneLauncher = hardwareMap.servo.get("droneLaunchServo");
         droneLauncher.setPosition(Constants.DRONE_START_POSITION);
+
         //Sets up dead wheels
         SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
-        Pose2d startPose = new Pose2d(-38.25, -63.75, Math.toRadians(90));
+        Pose2d startPose = new Pose2d(-38.25, 63.75, Math.toRadians(270));
         drive.setPoseEstimate(startPose);
 
         waitForStart();
 
         TrajectorySequence approachSpikeMarks = drive.trajectorySequenceBuilder(startPose)
                 //Move into spike mark area
-                .splineToLinearHeading(new Pose2d(-35, -30.25, startPose.getHeading()), startPose.getHeading())
+                .splineToLinearHeading(new Pose2d(-35, 30.25, startPose.getHeading()), startPose.getHeading())
                 //Determine where prop is located
                 .addDisplacementMarker(() -> {
                     if (leftSensor.getDistance(DistanceUnit.INCH) < 5) {
-                        selectPose = new Pose2d(-39, -31.25, Math.toRadians(180));
+                        selectPose = new Pose2d(-32, 31.25, Math.toRadians(0));
                         selectTurn = 90;
-                        boardOffset = 5.75;
-                        waitTime = 3.25;
-                    } else if (rightSensor.getDistance(DistanceUnit.INCH) < 5) {
-                        selectPose = new Pose2d(-32, -31.25,0);
-                        selectTurn = -90;
-                        boardOffset = -5.75;
+                        boardOffset = 4.25;
                         waitTime = 0;
+                    } else if (rightSensor.getDistance(DistanceUnit.INCH) < 5) {
+                        selectPose = new Pose2d(-39, 31.25,Math.toRadians(180));
+                        selectTurn = -90;
+                        boardOffset = -4.25;
+                        waitTime = 3.25;
                     } else {
-                        selectPose = new Pose2d(-35.5, -31.5, Math.toRadians(90));
+                        selectPose = new Pose2d(-35.5, 31.5, Math.toRadians(270));
                         selectTurn = 0;
-                        boardOffset = 0;
+                        boardOffset = 1.5;
                         waitTime = 2;
                     }
                 })
@@ -107,29 +110,29 @@ public class OldRedLeftAuto extends LinearOpMode {
                 //Back away from purple pixel
                 .back(4.5)
                 //Move back to safe turning position
-                .lineToLinearHeading(new Pose2d(-37, -48, selectPose.getHeading()))
+                .lineToLinearHeading(new Pose2d(-37, 48, selectPose.getHeading()))
                 //Turn towards back wall
-                .turn(Math.toRadians(179.999)-selectPose.getHeading())
+                .turn(Math.toRadians(180.001)-selectPose.getHeading())
                 //Drive towards back wall
-                .lineToLinearHeading(new Pose2d(-58,-48, Math.toRadians(180)))
+                .lineToLinearHeading(new Pose2d(-58,48, Math.toRadians(180)))
                 //Turn towards opposing side
-                .turn(Math.toRadians(-90))
+                .turn(Math.toRadians(90))
                 //Drive towards opposing side
-                .lineToLinearHeading(new Pose2d(-58,-9, Math.toRadians(90)))
+                .lineToLinearHeading(new Pose2d(-58,9, Math.toRadians(270)))
                 //Turn towards center of field
-                .turn(Math.toRadians(-90))
+                .turn(Math.toRadians(90))
                 //Drive towards center of field
-                .lineToLinearHeading(new Pose2d(15,-9,0))
+                .lineToLinearHeading(new Pose2d(15,9,0))
                 //Wait for teammate to finish
                 .waitSeconds(waitTime)
                 //Drive to board
-                .splineToLinearHeading(new Pose2d(39,-31 + boardOffset,0),Math.toRadians(-90))
+                .splineToLinearHeading(new Pose2d(39,38.5 + boardOffset,0),Math.toRadians(90))
                 .build();
 
         drive.followTrajectorySequence(placePurplePixel);
 
         //Raise slide and arm
-        slidePos = 600;
+        slidePos = 320;
         armUp = true;
         while ((slideMotor1.getCurrentPosition() < slidePos - 30 || rampPos < Constants.ARM_UP_POS - 0.001) && opModeIsActive()) {
             arm.setPosition(ArmRamp.Ramp(Constants.ARM_DOWN_POS, Constants.ARM_UP_POS, armUp));
@@ -182,6 +185,6 @@ public class OldRedLeftAuto extends LinearOpMode {
         slideMotor2.setPower(0);
 
         Constants.autoEndPose = dropOnBoard.end();
-        Constants.blueAuto = false;
+        Constants.blueAuto = true;
     }
 }
