+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/old/OldRedLeftAuto_Lower.java
@@ -1,7 +1,6 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.auto.old;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
-import com.acmerobotics.roadrunner.geometry.Vector2d;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
@@ -10,12 +9,14 @@ import com.qualcomm.robotcore.hardware.DistanceSensor;
 import com.qualcomm.robotcore.hardware.Servo;
 
 import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
+import org.firstinspires.ftc.teamcode.functions.ArmRamp;
+import org.firstinspires.ftc.teamcode.functions.Constants;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 
-@Autonomous(name="Blue Left Auto", group = "Centerstage Autonomous", preselectTeleOp = "RobotController")
+@Autonomous(name="Old Red Left Auto Lower", group = "Centerstage Autonomous", preselectTeleOp = "RobotController")
 @Disabled
-public class BlueLeftAuto extends LinearOpMode {
+public class OldRedLeftAuto_Lower extends LinearOpMode {
     //Sets up motors and variables
     DistanceSensor leftSensor;
     DistanceSensor rightSensor;
@@ -32,23 +33,23 @@ public class BlueLeftAuto extends LinearOpMode {
     Servo arm;
     double slidePosError;
     double slidePower;
-
+    double waitTime;
     Servo droneLauncher;
 
     @Override
     public void runOpMode() throws InterruptedException
     {
-        //Sets up sensors
+        //Sets sensors
         leftSensor = hardwareMap.get(DistanceSensor.class, "checkLeft");
         rightSensor = hardwareMap.get(DistanceSensor.class, "checkRight");
-        //Sets up servos
+        //Sets grippers and other servos
         leftGripper = hardwareMap.servo.get("leftGripperServo");
         rightGripper = hardwareMap.servo.get("rightGripperServo");
 
         leftGripper.setPosition(Constants.GRIPPER_LEFT_CLOSE_POSITION);
         rightGripper.setPosition(Constants.GRIPPER_RIGHT_CLOSE_POSITION);
         arm = hardwareMap.servo.get("armServo");
-        //Sets up slides
+        //Sets slide motors
         slideMotor1 = hardwareMap.dcMotor.get("leftSlideMotor");
         slideMotor2 = hardwareMap.dcMotor.get("rightSlideMotor");
 
@@ -64,30 +65,33 @@ public class BlueLeftAuto extends LinearOpMode {
 
         droneLauncher = hardwareMap.servo.get("droneLaunchServo");
         droneLauncher.setPosition(Constants.DRONE_START_POSITION);
-
+        //Sets up dead wheels
         SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
-        Pose2d startPose = new Pose2d(14.5, 63.75, Math.toRadians(-90));
+        Pose2d startPose = new Pose2d(-38.25, -63.75, Math.toRadians(90));
         drive.setPoseEstimate(startPose);
 
         waitForStart();
 
         TrajectorySequence approachSpikeMarks = drive.trajectorySequenceBuilder(startPose)
                 //Move into spike mark area
-                .splineToLinearHeading(new Pose2d(11.75, 30.25, startPose.getHeading()), startPose.getHeading())
+                .splineToLinearHeading(new Pose2d(-35, -30.25, startPose.getHeading()), startPose.getHeading())
                 //Determine where prop is located
                 .addDisplacementMarker(() -> {
                     if (leftSensor.getDistance(DistanceUnit.INCH) < 5) {
-                        selectPose = new Pose2d(15.75, 31.25, Math.toRadians(0));
+                        selectPose = new Pose2d(-39, -31.25, Math.toRadians(180));
                         selectTurn = 90;
                         boardOffset = 4.25;
+                        waitTime = 3.25;
                     } else if (rightSensor.getDistance(DistanceUnit.INCH) < 5) {
-                        selectPose = new Pose2d(7.75, 31.25, Math.toRadians(180));
+                        selectPose = new Pose2d(-32, -31.25,0);
                         selectTurn = -90;
                         boardOffset = -4.25;
+                        waitTime = 0;
                     } else {
-                        selectPose = new Pose2d(11.75, 31.5, Math.toRadians(-90));
+                        selectPose = new Pose2d(-35.5, -31.5, Math.toRadians(90));
                         selectTurn = 0;
-                        boardOffset = 1.5;
+                        boardOffset = -1.5;
+                        waitTime = 2;
                     }
                 })
                 .build();
@@ -105,11 +109,23 @@ public class BlueLeftAuto extends LinearOpMode {
                 //Back away from purple pixel
                 .back(4.5)
                 //Move back to safe turning position
-                .lineToLinearHeading(new Pose2d(13.75, 50, selectPose.getHeading()))
+                .lineToLinearHeading(new Pose2d(-37, -48, selectPose.getHeading()))
                 //Turn towards back wall
-                .turn(Math.toRadians(0.001)-selectPose.getHeading())
-                //Drives to board
-                .splineToLinearHeading(new Pose2d(39,37 + boardOffset,0),Math.toRadians(-90))
+                .turn(Math.toRadians(179.999)-selectPose.getHeading())
+                //Drive towards back wall
+                .lineToLinearHeading(new Pose2d(-58,-48, Math.toRadians(180)))
+                //Turn towards opposing side
+                .turn(Math.toRadians(-90))
+                //Drive towards opposing side
+                .lineToLinearHeading(new Pose2d(-58,-9, Math.toRadians(90)))
+                //Turn towards center of field
+                .turn(Math.toRadians(-90))
+                //Drive towards center of field
+                .lineToLinearHeading(new Pose2d(15,-9,0))
+                //Wait for teammate to finish
+                .waitSeconds(waitTime)
+                //Drive to board
+                .splineToLinearHeading(new Pose2d(39,-31 + boardOffset,0),Math.toRadians(-90))
                 .build();
 
         drive.followTrajectorySequence(placePurplePixel);
@@ -125,7 +141,6 @@ public class BlueLeftAuto extends LinearOpMode {
 
             slideMotor1.setPower(slidePower);
             slideMotor2.setPower(slidePower);
-
             telemetry.addData("Slide Pos", slideMotor1.getCurrentPosition());
             telemetry.addData("Arm Pos", arm.getPosition());
             telemetry.update();
@@ -168,15 +183,7 @@ public class BlueLeftAuto extends LinearOpMode {
         slideMotor1.setPower(0);
         slideMotor2.setPower(0);
 
-        TrajectorySequence moveToCorner = drive.trajectorySequenceBuilder(dropOnBoard.end())
-                //Moves to back right corner
-                .lineToConstantHeading(new Vector2d(44.5, 60))
-                //Scoots forward to get out of the way
-                .forward(10)
-                .build();
-        drive.followTrajectorySequence(moveToCorner);
-
-        Constants.autoEndPose = moveToCorner.end();
-        Constants.blueAuto = true;
+        Constants.autoEndPose = dropOnBoard.end();
+        Constants.blueAuto = false;
     }
 }
