+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/red/RedLeftAutoVisionPickUpWhite.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.auto.red;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
@@ -13,6 +13,8 @@ import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
 import org.firstinspires.ftc.robotcore.external.hardware.camera.controls.ExposureControl;
 import org.firstinspires.ftc.robotcore.external.hardware.camera.controls.GainControl;
 import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
+import org.firstinspires.ftc.teamcode.functions.ArmRamp;
+import org.firstinspires.ftc.teamcode.functions.Constants;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 import org.firstinspires.ftc.vision.VisionPortal;
@@ -22,8 +24,8 @@ import org.firstinspires.ftc.vision.apriltag.AprilTagProcessor;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-@Autonomous(name = "Blue Right Auto Vision With White", group = "Centerstage Autonomous Blue", preselectTeleOp = "RobotController")
-public class BlueRightAutoVisionPickUpWhite extends LinearOpMode {
+@Autonomous(name="Red Left Auto Vision With White", group = "Centerstage Autonomous Red", preselectTeleOp = "RobotController")
+public class RedLeftAutoVisionPickUpWhite extends LinearOpMode {
     //Sets up motors and variables
     DistanceSensor leftSensor;
     DistanceSensor rightSensor;
@@ -86,30 +88,30 @@ public class BlueRightAutoVisionPickUpWhite extends LinearOpMode {
 
         //Sets up dead wheels
         SampleMecanumDrive drive = new SampleMecanumDrive(hardwareMap);
-        Pose2d startPose = new Pose2d(-38.25, 63.75, Math.toRadians(270));
+        Pose2d startPose = new Pose2d(-38.25, -63.75, Math.toRadians(90));
         drive.setPoseEstimate(startPose);
 
-        Constants.blueAuto = true;
+        Constants.blueAuto = false;
 
         waitForStart();
 
         //Move into spike mark area
         Trajectory approachSpikeMarks = drive.trajectoryBuilder(startPose)
-                .splineToLinearHeading(new Pose2d(-35, 40.25, startPose.getHeading()), startPose.getHeading())
-                .lineToLinearHeading(new Pose2d(-35, 30.25, startPose.getHeading()))
+                .splineToLinearHeading(new Pose2d(-35, -40.25, startPose.getHeading()), startPose.getHeading())
+                .lineToLinearHeading(new Pose2d(-35, -30.25, startPose.getHeading()))
                 .build();
         drive.followTrajectory(approachSpikeMarks);
 
         TrajectorySequence placePurplePixel;
 
         //Detect prop
-        if (rightSensor.getDistance(DistanceUnit.INCH) < 5) {
-            selectPose = new Pose2d(-38.5, 31.25, Math.toRadians(180));
-            selectTurn = -90;
-            boardOffset = -6;
+        if (leftSensor.getDistance(DistanceUnit.INCH) < 5) {
+            selectPose = new Pose2d(-38.5, -31.25, Math.toRadians(180));
+            selectTurn = 90;
+            boardOffset = 6;
             zeOffset = 0;
-            waitTime = 0;
-            desiredTagId = 3;
+            waitTime = 2;
+            desiredTagId = 4;
             placePurplePixel = drive.trajectorySequenceBuilder(approachSpikeMarks.end())
                     //Turn to proper spike mark
                     .turn(Math.toRadians(selectTurn))
@@ -120,17 +122,15 @@ public class BlueRightAutoVisionPickUpWhite extends LinearOpMode {
                         leftGripper.setPosition(Constants.GRIPPER_LEFT_OPEN_POSITION);
                     })
                     .back(5)
-                    .turn(Math.toRadians(90))
-                    .lineToLinearHeading(new Pose2d(-37, 11.5, Math.toRadians(270)))
-                    .turn(Math.toRadians(-90))
+                    .lineToLinearHeading(new Pose2d(-37, -15, Math.toRadians(180)))
                     .build();
-        } else if (leftSensor.getDistance(DistanceUnit.INCH) < 5) {
-            selectPose = new Pose2d(-32.5, 31.25,0);
-            selectTurn = 90;
-            boardOffset = 6;
+        } else if (rightSensor.getDistance(DistanceUnit.INCH) < 5) {
+            selectPose = new Pose2d(-32.5, -31.25,0);
+            selectTurn = -90;
+            boardOffset = -6;
             zeOffset = 0;
             waitTime = 0;
-            desiredTagId = 1;
+            desiredTagId = 6;
             placePurplePixel = drive.trajectorySequenceBuilder(approachSpikeMarks.end())
                     //Turn to proper spike mark
                     .turn(Math.toRadians(selectTurn))
@@ -141,30 +141,31 @@ public class BlueRightAutoVisionPickUpWhite extends LinearOpMode {
                         leftGripper.setPosition(Constants.GRIPPER_LEFT_OPEN_POSITION);
                     })
                     .back(7)
-                    .turn(Math.toRadians(-180))
+                    .turn(Math.toRadians(90))
+                    .lineToLinearHeading(new Pose2d(-38, -11.5, Math.toRadians(90)))
+                    .turn(Math.toRadians(90))
                     .build();
         } else {
-            selectPose = new Pose2d(-35.5, 31.5, Math.toRadians(90));
-            selectTurn = -180;
+            selectPose = new Pose2d(-35.5, -31.5, Math.toRadians(90));
+            selectTurn = 180;
             boardOffset = 0;
             zeOffset = 0;
             waitTime = 1;
-            desiredTagId = 2;
+            desiredTagId = 5;
             placePurplePixel = drive.trajectorySequenceBuilder(approachSpikeMarks.end())
                     //Turn to proper spike mark
                     .turn(Math.toRadians(selectTurn))
                     //Drive up to proper spike mark
-                    .lineToLinearHeading(new Pose2d(-36, 16, Math.toRadians(90)))
+                    .lineToLinearHeading(new Pose2d(-40, -16, Math.toRadians(270.001)))
                     .addDisplacementMarker(() -> {
                         leftGripper.setPosition(Constants.GRIPPER_LEFT_OPEN_POSITION);
                     })
                     .back(8)
-                    .turn(Math.toRadians(90))
+                    .turn(Math.toRadians(-90))
                     .build();
         }
 
         drive.followTrajectorySequence(placePurplePixel);
-
         //TODO: Teleport to pick up white pixel
         //Raises arm
         slidePos = Constants.SLIDE_STACK_POS;
@@ -188,10 +189,10 @@ public class BlueRightAutoVisionPickUpWhite extends LinearOpMode {
         }
         slideMotor1.setPower(0);
         slideMotor2.setPower(0);
-        //  sleep(2147483647);
+      //  sleep(2147483647);
         TrajectorySequence actuallyPickUpWhitePixel = drive.trajectorySequenceBuilder(placePurplePixel.end())
                 //Drives into pixel stack
-                .splineToLinearHeading(new Pose2d(-60.5, 13.25, Math.toRadians(180)), Math.toRadians(180))
+                .splineToLinearHeading(new Pose2d(-60.5, -9.5, Math.toRadians(180)), Math.toRadians(180))
                 .build();
         drive.followTrajectorySequence(actuallyPickUpWhitePixel);
 
@@ -202,7 +203,7 @@ public class BlueRightAutoVisionPickUpWhite extends LinearOpMode {
 
         TrajectorySequence backUp = drive.trajectorySequenceBuilder(actuallyPickUpWhitePixel.end())
                 //Backs up a little bit to get room
-                .lineToLinearHeading(new Pose2d(-38, 11.5, Math.toRadians(180)))
+                .lineToLinearHeading(new Pose2d(-38, -11.5, Math.toRadians(180)))
                 .build();
         drive.followTrajectorySequence(backUp);
         //Lowers arm
@@ -243,20 +244,19 @@ public class BlueRightAutoVisionPickUpWhite extends LinearOpMode {
         sleep(0);
 
         TrajectorySequence holdWhitePixel = drive.trajectorySequenceBuilder(goForward.end())
-                .turn(Math.toRadians(180))
+                .turn(Math.toRadians(-180))
                 .build();
 
         drive.followTrajectorySequence(holdWhitePixel);
-
 
         //Waits for teammate to finish
         sleep(waitTime * 1000);
 
         TrajectorySequence driveToCenterOfField = drive.trajectorySequenceBuilder(holdWhitePixel.end())
                 //Drive towards center of field
-                .lineToLinearHeading(new Pose2d(15,9,0))
+                .lineToLinearHeading(new Pose2d(15,-9,0))
                 //Drive to board
-                .splineToLinearHeading(new Pose2d(36,35.5 + boardOffset,0), Math.toRadians(90))
+                .splineToLinearHeading(new Pose2d(36,-34 + boardOffset,0), Math.toRadians(-90))
                 .build();
 
         drive.followTrajectorySequence(driveToCenterOfField);
@@ -264,8 +264,8 @@ public class BlueRightAutoVisionPickUpWhite extends LinearOpMode {
         //Raise slide and arm
         slidePos = 550;
         armUp = true;
-        lastSlidePower = 0;
-        slidePowerIncr = 0.1;
+         lastSlidePower = 0;
+         slidePowerIncr = 0.1;
         while ((slideMotor1.getCurrentPosition() < slidePos - 30 || ArmRamp.rampPos > Constants.ARM_UP_POS + 0.001) && opModeIsActive()) {
             arm.setPosition(ArmRamp.Ramp(Constants.ARM_DOWN_POS, Constants.ARM_UP_POS, armUp));
             slidePosError = slidePos - slideMotor1.getCurrentPosition();
@@ -282,7 +282,6 @@ public class BlueRightAutoVisionPickUpWhite extends LinearOpMode {
             telemetry.addData("Slide Pos", slideMotor1.getCurrentPosition());
             telemetry.addData("Arm Pos", arm.getPosition());
             telemetry.update();
-
             lastSlidePower = slideMotor1.getPower();
         }
         slideMotor1.setPower(0);
@@ -340,10 +339,10 @@ public class BlueRightAutoVisionPickUpWhite extends LinearOpMode {
         }
 
         TrajectorySequence moveWhitePixel;
-        if(desiredTagId == 1){
+        if(desiredTagId == 4){
             //Drives forward to board
             Trajectory driveToBoard = drive.trajectoryBuilder(driveToCenterOfField.end())
-                    .lineToLinearHeading(new Pose2d(48 - ye, 32.5 - 12 /*# is white pixel offset*/ + boardOffset - ze, Math.toRadians(0 + pe)))
+                    .lineToLinearHeading(new Pose2d(48 - ye, -32.5 - 15 /*# is white pixel offset*/ + boardOffset - ze, Math.toRadians(0 + pe)))
                     .build();
             drive.followTrajectory(driveToBoard);
             leftGripper.setPosition(Constants.GRIPPER_LEFT_OPEN_POSITION);
@@ -359,7 +358,7 @@ public class BlueRightAutoVisionPickUpWhite extends LinearOpMode {
         else {
             //Drives forward to board
             TrajectorySequence driveToBoard = drive.trajectorySequenceBuilder(driveToCenterOfField.end())
-                    .lineToLinearHeading(new Pose2d(48.5 - ye, 35.5 + boardOffset - ze, Math.toRadians(0 + pe)))
+                    .lineToLinearHeading(new Pose2d(48 - ye, -32.5 + boardOffset - ze, Math.toRadians(0 + pe)))
                     .build();
             drive.followTrajectorySequence(driveToBoard);
 
@@ -372,9 +371,9 @@ public class BlueRightAutoVisionPickUpWhite extends LinearOpMode {
             drive.followTrajectorySequence(moveWhitePixel);
             leftGripper.setPosition(Constants.GRIPPER_LEFT_OPEN_POSITION);
         }
+
         //Waits for pixel to drop
         sleep(0);
-
         //Drives back from board
         Trajectory driveBackFromBoard = drive.trajectoryBuilder(moveWhitePixel.end())
                 .back(5)
