+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/functions/ArmRamp.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.functions;
 
 import com.qualcomm.robotcore.util.Range;
